Add files via upload
These are the new "User" instances. I'm submitting an extra pull request for the updated tests that change all the Users to Students. They can't be called User or else they don't work with SQL because of a pre-existing instance called user in SQL. This uses a JPA repository with an apache derby database to store things. While I didn't get to the UI part yet, you can use the free app called POSTMAN to test it.